Youtube query / config popout
Youtube api request seems to be working for the most part, but some queries are weird -> example 'cop' returns a vid fine but 'cops' returns undefined to server. It may be because the first result could potentially be a user channel rather than an actual video that returns.

Majority of this commit however focused on frontend changes with a config popout on bottom right using the ascii gear icon.  Implemented two checkboxes that disable images or disable videos people post.  Would also like another button next to it for immediate NSFW stuff to be canned as fast as the user can :P