code review is in need :hammer:
So this commit is pretty big.

I restrapped chat as a class. My thinking was to reuse the gui I've created in additional message windows to use for single one on one chat.  Clicking on a user spawns the new window, so thats covered, but is this the right way to go?