save drawing implementation (has bug)
unfortunately this tool has to be clicked twice in order to save.  i dont know the reason for this, but i have a feeling its because ev.target.download is called before the href is actually set.  the reason for toBlob() instead of getImageData() is because if a user posts a large image, the anchor fails to get the full base64 encoding and the file just shits itself unfortunately, which is the reason I'm converting it to a blob. could use help w/ this.